Add files via upload
This code presents the reconstruction of Figures 1-5, as referenced in the study "Simulation of a Brownian Particle
in an Optical Trap," available at "https://pubs.aip.org/aapt/ajp/article-abstract/81/3/224/1032381/Simulation-of-a-Brownian-particle-in-an-optical?redirectedFrom=fulltext". The reconstruction process utilized Python programming language. Efforts were made to ensure result reproducibility by implementing a 'seed' value in most figure 
 generation scripts.This ensures precise replication of the figures through code execution.